SVG app icon: drop broken <use> entries href'ing to non-existing id
"facesize" is an id for a helper shape in the original KMahjongg tileset
theme Jade, which seems to have slipped into the design of the app icon
when the "DRAGON_3" tile face was used as material.

QtSvg fills the log with "qt.svg: link #facesize is undefined!" due to
this, better avoided.

(cherry picked from commit 3343dbaf7d0842dcff2146a299072dd0c3e9fdea)